releasetools: Rebuild recovery when signing target images
When signing a build, the recovery can be built twice, once when signing
and again when adding the image to target files.

This is usally not a problem for deterministic signatures but a custom
signing method might include a time-based component which affects
checksums of the generated images. This means that the recovery patches
(i.e. install-recovery) will fail to checksum.

This patch makes adding images to target files, when invoked from
sign_target_files_apks, always rebuild the recovery patch with the newly
generated images so that the checksums are consistent irrespective of
potential time components.

Change-Id: I575c69dc96a6d4d69b3ea9d037f5f13c97afa87e